Change mesh primitive render count to take a series of ranges - only works for non-indexed VBOs

+++ mpp-helper/include/mpp/helper/QuadBatchRenderer.h
@@ -588,7 +588,7 @@ namespace mpp
 				}
 
 				mUniforms->setUniform("DIFFUSE", glm::vec4(1.0f, 1.0f, 1.0f, 1.0f));
-				mParams->setModelPointSize(params.getWidth());
+				mParams->setModelPointSize((float)params.getWidth());
 			}
 
 			virtual ~QuadBatchRenderer()
